+++ resale/src/main/java/com/goti/resale/scheduler/ResaleScheduler.java
@@ -1,4 +1,4 @@
-package com.goti.resale.utils;
+package com.goti.resale.scheduler;
 
 import java.time.LocalDateTime;
 import java.util.List;
